+++ src/data_sources/http/services/update_one/mod.rs
@@ -8,23 +8,23 @@ use crate::data_sources::http::filter::HttpDataSourceFilter;
 use super::Services;
 
 impl Services {
-    pub async fn find_one(
+    pub async fn update_one(
         client: Client,
         filter: HttpDataSourceFilter,
     ) -> Result<JsonValue, async_graphql::Error> {
-        info!("Executing Find One - HTTP Data Source");
+        info!("Executing Update One - HTTP Data Source");
 
         let response = Services::request(client, filter).await?;
-        debug!("Response Received: {:?}", response);
+        debug!("Response: {:?}", response);
 
         let json = json::parse(&response);
-        debug!("Response in JSON: {:?}", json);
+        debug!("JSON: {:?}", json);
 
         let res = match json {
-            Ok(res) => Ok(res),
+            Ok(json) => Ok(json),
             Err(error) => {
                 debug!("{:?}", error);
-                Err(Error::new("HTTP Find One Failed")
+                Err(Error::new("HTTP Update One Failed")
                     .extend_with(|err, e| e.set("details", err.message.as_str())))
             }
         };