+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - HTTP Data Source Added with Find One, Create One, and Find Many Resolvers
 - Optional property `exclude_from_input` is now a part of a Field definition.
+- Update One Resolver enables the ability to update mongo or put/patch http entities.
 
 ### Fixed
 - The `_id` property will no longer be removed automatically from input generation.